requirements.txt
Flask==1.1.2
calplot==0.1.7.3
petl==1.7.8
requests==2.25.1
plotly==4.14.3
numpy==1.19.3
sweetviz==2.1.3
wordcloud==1.8.1
chart_studio==1.1.0
seaborn==0.11.0
bokeh==2.3.0
streamlit==1.5.1
matplotlib==3.5.1
squarify==0.4.3
plotnine==0.8.0
cufflinks==0.17.3
ipywidgets==7.6.3
altair==4.1.0
bqplot==0.12.29
holoviews==1.14.4
panel==0.11.3
pandas==1.1.4
ipython==8.0.1
Pillow==9.0.1
scikit_learn==1.0.2
pydeck==0.6.1
spacy>=2.2.0,<3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm